+++ tests/layers/block/test_block_transformer.py
@@ -29,11 +29,14 @@ class TestTransformerProcessorBlock:
         num_heads=st.integers(min_value=1, max_value=10),
         activation=st.sampled_from(["ReLU", "GELU", "Tanh"]),
         window_size=st.integers(min_value=1, max_value=512),
+        dropout_p=st.floats(min_value=0.0, max_value=1.0),
     )
     @settings(max_examples=10)
-    def test_init(self, factor_attention_heads, hidden_dim, num_heads, activation, window_size):
+    def test_init(self, factor_attention_heads, hidden_dim, num_heads, activation, window_size, dropout_p):
         num_channels = num_heads * factor_attention_heads
-        block = TransformerProcessorBlock(num_channels, hidden_dim, num_heads, activation, window_size)
+        block = TransformerProcessorBlock(
+            num_channels, hidden_dim, num_heads, activation, window_size, dropout_p=dropout_p
+        )
         assert isinstance(block, TransformerProcessorBlock)
 
         assert isinstance(block.layer_norm1, nn.LayerNorm)
@@ -49,6 +52,7 @@ class TestTransformerProcessorBlock:
         window_size=st.integers(min_value=1, max_value=512),
         shapes=st.lists(st.integers(min_value=1, max_value=10), min_size=3, max_size=3),
         batch_size=st.integers(min_value=1, max_value=40),
+        dropout_p=st.floats(min_value=0.0, max_value=1.0),
     )
     @settings(max_examples=10)
     def test_forward_output(
@@ -60,9 +64,12 @@ class TestTransformerProcessorBlock:
         window_size,
         shapes,
         batch_size,
+        dropout_p,
     ):
         num_channels = num_heads * factor_attention_heads
-        block = TransformerProcessorBlock(num_channels, hidden_dim, num_heads, activation, window_size)
+        block = TransformerProcessorBlock(
+            num_channels, hidden_dim, num_heads, activation, window_size, dropout_p=dropout_p
+        )
 
         x = torch.randn((batch_size, num_channels))